… CHANGELOG.md
… tempodb/encoding/vparquet4/block_autocomplete.go